archive/v1_2025/src/backend/requirements.txt
# ELWOSA Backend Requirements
# Python 3.10+

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1
sqlalchemy==2.0.23

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# WebSocket Support
websockets==12.0
python-socketio==5.10.0

# Redis Cache
redis==5.0.1
aioredis==2.0.1

# API Documentation
pydantic==2.5.0
pydantic-settings==2.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging & Monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# AI Integration (Optional)
openai==1.3.5
langchain==0.0.340